introduce make test-heavy
Move the dependence of zq-sample-data to its own target, 'make
test-heavy'. The name heavy is chosen to imply dependence on a larger
dataset. There is other testing besides markdown verfication that can
use this.

A separate make target requires a separate CircleCI step.
